Update launch script to better handle tty switching and add some comments.
onHome will now keep an app running on it's tty instead of possibly
killing it. SDL apps are trickier than console apps since they grab
the tty and won't let us chvt out. The script will attempt to release
SDL's hold on the tty with termfix and then chvt. This is a temporary
solution until we find a way to tell SDL to Iconify the window so we
can chvt. Some SDL apps do not work with onHome, but you can still use
ALT-RIGHT/LEFT on the Zipit to switch vt's.